+++ src/sinks/gcp/mod.rs
@@ -6,7 +6,7 @@ use vector_config::configurable_component;
 pub mod chronicle_unstructured;
 pub mod cloud_storage;
 pub mod pubsub;
-pub mod stackdriver_logs;
+pub mod stackdriver;
 pub mod stackdriver_metrics;
 
 /// A monitored resource.
