+++ src/sinks/gcp/stackdriver/logs/service.rs
@@ -1,0 +1,25 @@
+//! Service implementation for the `http` sink.
+
+use bytes::Bytes;
+use http::{Request, Uri};
+
+use crate::{gcp::GcpAuthenticator, sinks::util::http::HttpServiceRequestBuilder};
+
+#[derive(Debug, Clone)]
+pub(super) struct StackdriverLogsSinkRequestBuilder {
+    pub(super) uri: Uri,
+    pub(super) auth: GcpAuthenticator,
+}
+
+impl HttpServiceRequestBuilder for StackdriverLogsSinkRequestBuilder {
+    fn build(&self, body: Bytes) -> Request<Bytes> {
+        let mut request = Request::post(self.uri.clone())
+            .header("Content-Type", "application/json")
+            .body(body)
+            .unwrap();
+
+        self.auth.apply(&mut request);
+
+        request
+    }
+}
