+++ src/sinks/gcp/stackdriver/mod.rs
@@ -1,0 +1,1 @@
+mod logs;
